__wt_epoch() should be identical between POSIX and Windows, and isn't. Pull the higher-level time functions out into a new file.
